requirements/dev.txt
-r common.txt
django-debug-toolbar<1.9
ipython<5.6
python-daemon<2.2